Move guide links and response buttons to separate components
